Final Changes (5 Files)
Changed about us css and html to fit the rest of the site stylistically, changed font and nav bar. Calculator JS is having an issue on my end, not sure what the issue is but @diabeatz96 fixed it. Muscles.js updated with functionality and working links. Changed readme slightly and revised certain things.